Add maps and metadata for JournalL3
Because of the use of a for-loop to iterate through the list of call number ranges, a virtual "shelf" that targets a single call number can be retrieved even if the value of the call number is between range_start and range_end of another shelf, so long as in map_it.json the single call number is defined first. If we move away from the for-loop (say, to NeDB), there is no guarantee that the single call number will be "hit" first, hence it will affect the ability to retrieve it.