Add intershelving for locations
Adds a new key intershelve_with in the data file to indicate that the items from that location are intershelved at the target location. The value of intershelve_with should be the key/index of the target location.

The functionality only supports intershelving between locations inside the same library.

A location that is retrieved by calling dbo.findLocationByCode() will contain a copy of the shelves from the intershelving location. On the other hand, the child locations inside a library returned by calling dbo.findLibraryByCode() do not get this copying.